moves files into new packages
