+++ src/main/java/de/rabea/gui/clickhandler/FullCellClickHandler.java
@@ -1,6 +1,10 @@
-package de.rabea.gui;
+package de.rabea.gui.clickhandler;
 
 import de.rabea.game.Board;
+import de.rabea.gui.ClickHandler;
+import de.rabea.gui.GuiApp;
+import de.rabea.gui.GuiPlayer;
+import de.rabea.gui.ViewUpdater;
 
 public class FullCellClickHandler implements ClickHandler {
