+++ src/main/java/de/rabea/gui/clickhandler/GameOptionsClickHandler.java
@@ -1,4 +1,7 @@
-package de.rabea.gui;
+package de.rabea.gui.clickhandler;
+
+import de.rabea.gui.ClickHandler;
+import de.rabea.gui.GuiApp;
 
 import static de.rabea.game.GameMode.GuiHumanVsComputer;
 import static de.rabea.game.GameMode.GuiHumanVsGuiHuman;
