+++ src/main/java/de/rabea/gui/view/BoardSizeView.java
@@ -1,5 +1,8 @@
-package de.rabea.gui;
+package de.rabea.gui.view;
 
+import de.rabea.gui.GuiApp;
+import de.rabea.gui.JavaFXButton;
+import de.rabea.gui.clickhandler.BoardSizeClickHandler;
 import javafx.scene.Parent;
 import javafx.scene.control.Label;
 import javafx.scene.layout.GridPane;
