+++ src/test/java/de/rabea/gui/GuiPlayerTest.java
@@ -5,9 +5,7 @@ import de.rabea.game.Mark;
 import org.junit.Test;
 
 import static de.rabea.game.Mark.O;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 public class GuiPlayerTest {
