+++ src/test/java/de/rabea/gui/clickhandler/EmptyCellClickHandlerTest.java
@@ -1,6 +1,9 @@
-package de.rabea.gui;
+package de.rabea.gui.clickhandler;
 
 import de.rabea.game.Board;
+import de.rabea.gui.GuiApp;
+import de.rabea.gui.GuiPlayer;
+import de.rabea.gui.ViewUpdater;
 import de.rabea.player.PlayerFactory;
 import org.junit.Test;
