+++ src/test/java/de/rabea/gui/clickhandler/FullCellClickHandlerTest.java
@@ -1,6 +1,9 @@
-package de.rabea.gui;
+package de.rabea.gui.clickhandler;
 
 import de.rabea.game.Board;
+import de.rabea.gui.GuiApp;
+import de.rabea.gui.GuiPlayer;
+import de.rabea.gui.ViewUpdaterSpy;
 import org.junit.Test;
 
 import static org.junit.Assert.assertTrue;
